refactor: 파일 및 코드 정리
