+++ app/(list)/_components/Gathering.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import ListContainer from '@/components/list/ListContainer';
 import GatheringHeader from './gatheringCard/GatheringHeader';
 import GatheringTabs from './gatheringCard/GatheringTabs';
-import GatheringCardList from './gatheringCard/GatheringCardList';
+import GatheringCList from './gatheringCard/GatheringList';
 
 export default function Gathering() {
   return (
     <ListContainer>
       <GatheringHeader />
       <GatheringTabs />
-      <GatheringCardList />
+      <GatheringCList />
     </ListContainer>
   );
 }
