+++ app/(list)/_components/gatheringCard/DeadlineBadge.tsx
@@ -16,7 +16,7 @@ const DeadlineBadge = ({ registrationEnd, roundedTopRight = true }: Props) => {
   // 24시간 이상 남으면 일수로, 그렇지 않으면 시간으로 표시
   let remainingTime;
   if (hoursDiff <= 0) {
-    remainingTime = '마감';
+    return null;
   } else if (hoursDiff <= 23) {
     remainingTime = `${hoursDiff}시간 남음`;
   } else {
